Expanded on message dictionary
Now has a list of responses for each message, added to instead of overwriting previously existing responses. This allows CHELSEA to have more flexibility in how she responds to the same messages. I decided to do this because I figured CHELSEA was losing valuable information by overwriting the previous responses, it's better that she tacks new ones onto a list instead.

Also, under the 'tools' directory is the script for converting the Version 0.12 'messageDictionary.json' memory file to the new one for Version 0.13+.