Fixed another 'what is/are' problem
Realized the 'what is/are' questions were being stored with the 'is/are' part intact. I didn't edit her memory containing those, but fixed it so in the future the 'is/are' gets replaced by either 'is' or 'are' from the captured group of the user_message, which makes more sense.