New version, new format for mem file, see conversion
If you have been using the 'BETA11' version, you can still use the memory files for 'BETA12', just use the conversion code that comes with 'BETA12'.
CHELSEA can now add more unanswered questions, in the form of 'why is/are', and look for a valid answer from the user.
Currently, the questions formed often don't follow proper grammar rules, so if one of those questions pop up, you can say:
"That's not proper grammar", "That is incorrect grammar", etc. to have CHELSEA delete the question.
This is a temporary solution to that problem, I would like to add some kind of code to detect and filter the incorrect grammar.